got rid of annoying warnings about comparison between int and size_t

+++ TrajectoryVisualizer/view/main_view.cpp
@@ -72,7 +72,7 @@ void MainView::setFirstTrajectory(const vector<QPixmap> &imgs, const vector<QPoi
 {
 
     scene.getFirstTrajectory().clear();
-    for (int i = 0; i < imgs.size(); i++)
+    for (size_t i = 0; i < imgs.size(); i++)
     {
         scene.getFirstTrajectory().pushBackFrame(imgs[i], center_coords_px[i], angles[i], meters_per_pixels[i], qualities[i]);
     }
@@ -83,7 +83,7 @@ void MainView::setSecondTrajectory(const vector<QPixmap> &imgs, const vector<QPo
                                    const vector<double> &qualities)
 {
     scene.getSecondTrajectory().clear();
-    for (int i = 0; i < imgs.size(); i++)
+    for (size_t i = 0; i < imgs.size(); i++)
     {
         scene.getSecondTrajectory().pushBackFrame(imgs[i], center_coords_px[i], angles[i], meters_per_pixels[i], qualities[i]);
     }
@@ -93,7 +93,7 @@ void MainView::setFirstKeyPoints(const vector<int> &frames_num, const vector<QPo
                                  const vector<double> &radius, const vector<QColor> &colors)
 {
     scene.getFirstTrajectory().clearKeyPoints();
-    for (int i = 0; i < frames_num.size(); i++)
+    for (size_t i = 0; i < frames_num.size(); i++)
     {
         scene.getFirstTrajectory().addKeyPoint(frames_num[i], centers_px[i], angles[i], radius[i], colors[i]);
     }
@@ -103,7 +103,7 @@ void MainView::setSecondKeyPoints(const vector<int> &frames_num, const vector<QP
                                  const vector<double> &radius, const vector<QColor> &colors)
 {
     scene.getSecondTrajectory().clearKeyPoints();
-    for (int i = 0; i < frames_num.size(); i++)
+    for (size_t i = 0; i < frames_num.size(); i++)
     {
         scene.getSecondTrajectory().addKeyPoint(frames_num[i], centers_px[i], angles[i], radius[i], colors[i]);
     }
@@ -123,7 +123,7 @@ void MainView::setMatches(const std::vector<std::vector<QPointF>> &trajectory_pt
                           const std::vector<std::vector<double>> &meters_per_pixels)
 {
     scene.getMatches().clear();
-    for (int i = 0; i < trajectory_pts.size(); i++)
+    for (size_t i = 0; i < trajectory_pts.size(); i++)
     {
         scene.getMatches().addLine(trajectory_pts[i], frames_center_on_map[i], angles[i], meters_per_pixels[i]);
     }
